mac/requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0
pandas>=2.0.0
primer3-py>=2.0.0
pyinstaller>=6.0.0
